requirements.txt
python-dateutil>=2.8.0
google-cloud-bigquery==3.11.4
google-auth>=2.23.0
requests==2.31.0
pandas==2.0.3
openai==0.27.8
playwright>=1.40.0
Pillow>=10.1.0
google-cloud-storage==2.10.0
httpx>=0.24.0
google-oauth2-tool>=0.0.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.11.2
webdriver-manager==3.8.6
psycopg2-binary==2.9.7